WineBook 0.0.0 integration

+++ build.sh
@@ -1,25 +1,36 @@
+FRAMEWORK=$1;
 BUILD_DIR="build"
-OUTPUT="${BUILD_DIR}/Debug-Universal"
-rm -rf "${BUILD_DIR}"
-mkdir -p "${BUILD_DIR}"
-rm -rf "${OUTPUT}"
-mkdir -p "${OUTPUT}"
-
-# clear cache
-rm -rf ~/Library/Developer/Xcode/DerivedData
-rm -rf "${BUILD_DIR}"
-mkdir -p "${BUILD_DIR}" 
-rm -rf "${OUTPUT}"
-mkdir -p "${OUTPUT}" 
+OUTPUT="${BUILD_DIR}/Debug-Universal" 
 
 # build
-xcodebuild \
-    -quiet \
-    -workspace "Forms.xcworkspace" \
-    -scheme "Forms-Universal-Framework" \
-    -destination generic/platform=iOS \
-    -destination "platform=iOS Simulator,name=iPhone 11,OS=latest" \
-    SYMROOT=$(PWD)/build/
+if [ -z "$FRAMEWORK" ]
+then
+    # clear cache for full build
+    rm -rf ~/Library/Developer/Xcode/DerivedData
+    rm -rf "${BUILD_DIR}"
+    mkdir -p "${BUILD_DIR}" 
+    mkdir -p "${BUILD_DIR}/Debug-Universal" 
+    rm -rf "${OUTPUT}"
+    mkdir -p "${OUTPUT}" 
+    xcodebuild \
+        -quiet \
+        -workspace "Forms.xcworkspace" \
+        -scheme "Forms-Universal-Framework" \
+        -destination generic/platform=iOS \
+        -destination "platform=iOS Simulator,name=iPhone 11,OS=latest" \
+        SYMROOT=$(PWD)/build/
+else
+    mkdir -p "${BUILD_DIR}"
+    mkdir -p "${BUILD_DIR}/Debug-Universal"
+    mkdir -p "${OUTPUT}"
+    xcodebuild \
+        -quiet \
+        -workspace "Forms.xcworkspace" \
+        -scheme "$FRAMEWORK" \
+        -destination generic/platform=iOS \
+        -destination "platform=iOS Simulator,name=iPhone 11,OS=latest" \
+        SYMROOT=$(PWD)/build/
+fi
 
 # define frameworks
 frameworks=(
@@ -37,6 +48,7 @@ frameworks=(
     FormsLogger
     FormsMock
     FormsNetworking
+    FormsNetworkingImage
     FormsNotifications
     FormsPermissions
     FormsPagerKit
@@ -46,15 +58,19 @@ frameworks=(
     FormsToastKit
     FormsTransitions
     FormsUtils
+    FormsUtilsUI
     FormsValidators
 ) 
 
 # merge architectures
 for framework in "${frameworks[@]}"
 do 
-    cp -R "${BUILD_DIR}/Debug-iphoneos/${framework}.framework" "${OUTPUT}/"
-    lipo -create -output "${OUTPUT}/${framework}.framework/${framework}" "${BUILD_DIR}/Debug-iphoneos/${framework}.framework/${framework}" "${BUILD_DIR}/Debug-iphonesimulator/${framework}.framework/${framework}"
-    cp -r "${BUILD_DIR}/Debug-iphonesimulator/${framework}.framework/Modules/${framework}.swiftmodule/" "${OUTPUT}/${framework}.framework/Modules/${framework}.swiftmodule"
+    if [ -z "$FRAMEWORK" ] || [ $framework = $FRAMEWORK ]
+    then
+        cp -R "${BUILD_DIR}/Debug-iphoneos/${framework}.framework" "${OUTPUT}/"
+        lipo -create -output "${OUTPUT}/${framework}.framework/${framework}" "${BUILD_DIR}/Debug-iphoneos/${framework}.framework/${framework}" "${BUILD_DIR}/Debug-iphonesimulator/${framework}.framework/${framework}"
+        cp -r "${BUILD_DIR}/Debug-iphonesimulator/${framework}.framework/Modules/${framework}.swiftmodule/" "${OUTPUT}/${framework}.framework/Modules/${framework}.swiftmodule"
+    fi
 done
 
 open "${OUTPUT}"
